Contain all the proxy fetching logic in useProxiedFileUrl react effect
